--- src/main/java/org/example/article/cli/config/CommandConfig.java
@@ -1,9 +1,0 @@
-package org.example.article.cli.config;
-
-import org.example.article.cli.controller.CommandController;
-
-public interface CommandConfig {
-
-    CommandController commandController();
-
-}
