+++ src/main/java/org/example/article/presentation/config/CommandConfig.java
@@ -1,0 +1,9 @@
+package org.example.article.presentation.config;
+
+import org.example.article.presentation.controller.CommandController;
+
+public interface CommandConfig {
+
+    CommandController commandController();
+
+}
