+++ src/main/java/org/example/article/presentation/dto/request/IdDto.java
@@ -1,4 +1,4 @@
-package org.example.article.cli.dto.request;
+package org.example.article.presentation.dto.request;
 
 public class IdDto {
     private final String id;
